requirements.txt
duckduckgo-search
wikipedia
arxiv
langchain
python
langchain huggingface_hub
python-dotenv